+++ votesmart/__init__.py
@@ -1,3 +1,4 @@
 from .api import VoteSmartAPI
+from .exceptions import VotesmartApiError, VotesmartNotFoundError
 
-__all__ = ["VoteSmartAPI"]
+__all__ = ["VoteSmartAPI", "VotesmartApiError", "VotesmartNotFoundError"]